requirements.txt
elasticsearch>=8.13,<9
fastapi
uvicorn[standard]
langchain
python-multipart
uvicorn
pandas
python-docx
pydantic
openpyxl
camelot
